went back to Double

+++ src/main/java/com/aquila/chess/strategy/mcts/CacheValues.java
@@ -53,7 +53,7 @@ public class CacheValues {
         return ret;
     }
 
-    public synchronized CacheValue updateValueAndPolicies(long key, float value, float[] notNormalisedPolicies) {
+    public synchronized CacheValue updateValueAndPolicies(long key, double value, double[] notNormalisedPolicies) {
         CacheValue cacheValue = this.lruMap.get(key);
         if (cacheValue == null) {
             throw new RuntimeException("node for key:" + key + " not found");
@@ -68,7 +68,7 @@ public class CacheValues {
         private static final float NOT_INITIALIZED_VALUE = 0;
 
         private static final CacheValue getNotInitialized(final String label) {
-            return new CacheValue(NOT_INITIALIZED_VALUE, label, new float[PolicyUtils.MAX_POLICY_INDEX]);
+            return new CacheValue(NOT_INITIALIZED_VALUE, label, new double[PolicyUtils.MAX_POLICY_INDEX]);
         }
 
         @Setter
@@ -85,7 +85,7 @@ public class CacheValues {
 
         private int nbPropagate = 1;
 
-        private CacheValue(float value, String label, float[] policies) {
+        private CacheValue(double value, String label, double[] policies) {
             super(value, policies);
             this.label = label;
         }
@@ -102,13 +102,13 @@ public class CacheValues {
             this.type = CacheValueType.ROOT;
         }
 
-        public synchronized void normalise(float[] policies) {
+        public synchronized void normalise(double[] policies) {
             int[] indexes = PolicyUtils.getIndexesFilteredPolicies(node.getChildMoves());
             if (log.isDebugEnabled())
                 log.debug("NORMALIZED type:{} move.size:{} dirichlet:{}", this.type, node.getChildMoves().size(), node.isDirichlet());
             boolean isDirichlet =node.getState() == MCTSNode.State.ROOT;
             isDirichlet = MCTSStrategyConfig.isDirichlet(node.getMove()) && isDirichlet;
-            float[] normalisedPolicies = Utils.toDistribution(policies, indexes, isDirichlet);
+            double[] normalisedPolicies = Utils.toDistribution(policies, indexes, isDirichlet);
 //            if (Arrays.stream(normalisedPolicies).filter(policy -> Double.isNaN(policy)).count() > 0) {
 //                throw new RuntimeException("ERROR, some policy with NaN value");
 //            }
@@ -119,7 +119,7 @@ public class CacheValues {
             this.type = CacheValueType.LEAF;
         }
 
-        public OutputNN setTrueValuesAndPolicies(final float value, final float[] policies) {
+        public OutputNN setTrueValuesAndPolicies(final double value, final double[] policies) {
             this.value = value;
             this.policies = policies;
             log.debug("setTrueValuesAndPolicies({},{} {} {} ..)", value, policies[0], policies[1], policies[2]);
